backend/requirements.txt
dice<2.5
django-extensions<2.2
django-filter<2.2
django<2.3
djangorestframework<3.10
gunicorn
psycopg2-binary<2.9
